Hockey can delete players from T and C. Hockey can also delete Teams from T.

+++ clients/android/hockey/src/main/java/fit/cvut/org/cz/hockey/business/managers/PackagePlayerManager.java
@@ -12,9 +12,11 @@ import fit.cvut.org.cz.tmlibrary.business.entities.Player;
 import fit.cvut.org.cz.tmlibrary.business.entities.Team;
 import fit.cvut.org.cz.tmlibrary.business.entities.Tournament;
 import fit.cvut.org.cz.tmlibrary.business.interfaces.IPackagePlayerManager;
+import fit.cvut.org.cz.tmlibrary.data.entities.DParticipant;
 import fit.cvut.org.cz.tmlibrary.data.entities.DPlayer;
 import fit.cvut.org.cz.tmlibrary.data.entities.DStat;
 import fit.cvut.org.cz.tmlibrary.data.entities.DTeam;
+import fit.cvut.org.cz.tmlibrary.data.entities.DTournament;
 
 /**
  * Created by atgot_000 on 8. 4. 2016.
@@ -33,13 +35,41 @@ public class PackagePlayerManager implements IPackagePlayerManager {
 
     @Override
     public boolean deletePlayerFromCompetition(Context context, long playerId, long competitionId) {
+
+        ArrayList<DTournament> tournaments = DAOFactory.getInstance().tournamentDAO.getByCompetitionId( context, competitionId );
+
+        //Check if player is not in any tournament
+        for(DTournament tourn : tournaments ){
+            ArrayList<Long> playerIds = DAOFactory.getInstance().packagePlayerDAO.getPlayerIdsByTournament( context, tourn.getId() );
+            if( playerIds.contains( playerId ) ) return false;
+        }
+
         DAOFactory.getInstance().packagePlayerDAO.deletePlayerFromCompetition(context, playerId, competitionId);
+
         return true;
     }
 
     @Override
     public boolean deletePlayerFromTournament(Context context, long playerId, long tournamentId) {
+
+
+        ArrayList<Player> notInTeams = ManagerFactory.getInstance().packagePlayerManager.getPlayersNotInTeams( context, tournamentId );
+        boolean flag = false;
+        for( Player p : notInTeams ){
+            if( p.getId() == playerId ) {
+                flag = true;
+                break;
+            }
+        }
+        if( !flag ) return false;
+
+        ArrayList<DStat> tourStats = DAOFactory.getInstance().statDAO.getStatsByTournamentId( context, tournamentId );
+
+        for( DStat stat : tourStats ){
+            if( stat.getPlayerId() == playerId ) return false;
+        }
         DAOFactory.getInstance().packagePlayerDAO.deletePlayerFromTournament(context, playerId, tournamentId);
+
         return true;
     }
